applications: asset_tracker_v2: Prefix IMEI with "nrf-"
When passing in IMEI as the client ID to the nRF Cloud backend it
needs to be prefixed with "nrf-" by default. This was missed in
a previous PR that added support for passing in the client ID at
run time from the application. The nRF cloud backend does not
prefix client IDs that are passed in run time.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>