samples: nrf9160: lwm2m_client: Remove obsolete hotfix for buttons
This will remove a redundant hotfix for tracking the state of buttons in
lwm2m. Previously, zephyr would increment the Input Counter directly
meaning that tracking the value would not work. This hotfix recreated
the expected functionality as the fix in zephyr had not been merged yet.

The fix for this issue has been upmerged, so this hotfix is now
redundant and can be removed.

Signed-off-by: Markus Rekdal <markus.rekdal@nordicsemi.no>