dfu: multi-image package format and library
Implement Python script and CMake wrapper for
building a DFU Multi Image package consisting of
a CBOR-based header, followed by a list of selected update
images. If DFU_MULTI_IMAGE_PACKAGE_BUILD is set, build
the DFU Multi Image package with common update images, such
as app_update.bin, net_core_app_update.bin or MCUboot S0/S1.
Which images make it to the package can be configured using
separate Kconfig options.

Add library for downloading and parsing a DFU Multi Image
package. The library mimics the DFU Target library in terms
of API and can be easily integrated with the latter but is
not limited to that and a user may provide a custom image
writer for writing a specific firmware image included in the
package.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>