lib: sms: fix time zone handling for received SMSs
Time zone for received SMSs was decoded incorrectly. The parsed
time zone was also returned as full hours which is not sufficient
because there are time zones which use quarters of an hour.

Fixed the decoding and changed the time zone to be returned as
quarters of an hour as specified by 3GPP TS 23.040 which was
already referenced by the API.

NCSIDB-744

Signed-off-by: Tommi Kangas <tommi.kangas@nordicsemi.no>