doc: known-issues: add known issue for clock_gettime() API
Added a known issue for clock_gettime() API. The API starts
returning invalid time after one week has elapsed since the
system booted because of an overflow in the uptime conversion.
The issue is present in SDKs v1.9.0 and v1.9.1.

Added a known issue for the date_time library, which uses
the clock_gettime() API.

Signed-off-by: Tommi Kangas <tommi.kangas@nordicsemi.no>