doc: nrf: add a section with fota sample to nRF52 Series user guide
Added a new section for the SMP server samples to the nRF52 Series user
guide. This sample demonstrates how to perform FOTA upgrades in the nRF
Connect SDK.

The description will also be available as part of the nRF53 Series user
guide. It is included there automatically with the help of the FOTA
upgrades start and end tags.

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>