samples: ipc: Fix dts configuration
This commit fixes an issue of undefined reference to
some IPC nrfx functions. This issue happened when
ipc-service was compiled with icmsg backend.

The samples fixed in this commit were unnecessarily
redefining mbox node. In the defaul node definition
two compatibles are present:
"nordic,mbox-nrf-ipc" and "nordic,nrf-ipc"

Based on "nordic,nrf-ipc" driver that defines missing
functions was not included in the build.

This issue could be fixed by adding missing compatible,
however redefinition of mbox node was not necessary in
the first place, thus remove it.

With this patch, the default mbox node definition is used
and it uses both required compatibles.

Signed-off-by: Emil Obalski <emil.obalski@nordicsemi.no>