doc: Remove segger content from nRF connect SDK
The segger content would be removed from nRF connect SDK
documents and will be replaced with Visual code extension.

This PR removes the segger content and there would be following PRs
to add the VSC to the following documents:
•Modifying an application - NCSDK-14998
•Thingy:91 User guide - NCSDK-14991
•Working with nRF5340 DK - NCSDK-14992
•samples\nRF Secure Immutable Bootloader - NCSDK-14993

This PR will only be merged once the PRs for the above tasks
are merged.

Signed-off-by: divya pillai <divya.pillai@nordicsemi.no>