doc: Improve nRF Cloud device security docs
Add some cloud-related terms to the glossary.

Add a security section to the nRF Cloud user guide.

Reorganize the list of samples into MQTT, REST,
and other.

Add ug_nrf_cloud to the top level index so nRF Cloud
is not buried so deeply in the hierarchy and remove
from ext_comps.rst.

Signed-off-by: Pekka Niskanen <pekka.niskanen@nordicsemi.no>
Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>