applications: asset_tracker_v2: Change to deferred logging
Immediate log mode impacts the performace of the system because the
logging is being executed in context of the caller. Due to this some
time critical functions can fail if a lot of logging is enabled.
For instance when the debug configuration `overlay-debug.conf`
is enabled. This patch changes back to deferred logging to alleviate
this issue.

Deferred logging was previously reverted to immediate due to floats
not being properly logged. This has now been fixed.

In addition, logging for nRF Cloud A/P-GPS libraries are turned of to
reduce the stress on the logger. Without this config the user still get
enough information from the application to indicate reason of failure,
if any.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>