Bluetooth: Mesh: Gen Ponoff Srv status msg update
Alter status message responses to use setup srv instance when responding
to a SET message type (message associated with the Ponoff setup srv) and
the srv instance when responding to a GET message (message associated
with the Ponoff srv). This allows the Ponoff setup srv to respond to a
SET message from the client without a dependency on the Ponoff srv being
bound to the same appkey.

Signed-off-by: Anders Storrø <anders.storro@nordicsemi.no>