tfm: Add option to enable TF-M init of architecture state on boot
Add option for TF-M to enable initialization of architecture state on
boot. Enable this by default when MCUBoot has been selected as the
bootloader.
If CONFIG_MCUBOOT_CLEANUP_ARM_CORE is enabled in MCUBoot then this
option is not required, however in order to support already deployed
MCUBoot without this option enabled then this needs to be handled by
TF-M during boot to prevent a system crash.

NCSDK-15443

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>