applications: nrf_desktop: Avoid pulling in outside led_id enums
Doxygen does not distinguish enum's coming from different scopes.
If enum of the same name is declared in different scope (e.g. another
application), the doxygen will pull it in and if there are the same
names in these enums, the doxygen will report name conflicts.

Avoid it by using unique name for enum in nrf_desktop. This can
can be done painlessly as we don't reference explicitly name of
this enum in code.

Jira: NCSDK-14306

Signed-off-by: Pawel Dunaj <pawel.dunaj@nordicsemi.no>