samples: peripheral_hr_coded: central_hr_coded: Use valid configuration
The central_hr_coded and peripheral_hr_coded samples when
running with Zephyr LL cannot enable an advertising or
a scanning because for the Zephyr LL by default extended
advertising is disabled. This commit improves configuration
for those samples by enabling extended advertising in the BLE
controller and by adding a separated configuration for nRF53.

NCSDK-15527

Signed-off-by: Kamil Gawor <Kamil.Gawor@nordicsemi.no>