doc: improve Thread configuration instructions
This rearranges some of the Configuring Thread in NCS page and
and adds details to some sections to make it more clear what
the user needs to do, especially in regards to working with
our samples. Small edits are made to related pages as well.

Also takes out the information about pre-built libraries and
makes it its own page. In the same way that Thread 1.2 features
are not described on the configuration page, the pre-built
libraries don't need to be described here.

Signed-off-by: Wille Backman <wille.backman@nordicsemi.no>