github: disable tf-m oss-history check
Pull request #7933 is bringing in a lot of commits from upstream tf-m
that are not present in the revision of tf-m in zephyr/west.yml, and
the script is getting confused.

In particular, it thinks that all of the new upstream commits are
actually downstream out-of-tree commits. Then it tries to reapply them
onto the old downstream. This fails, and the action fails.

I'm not sure how to handle this case, so let's just disable the check
for tf-m for now.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>