application: serial_lte_modem: BUG-FIX enter data mode
The swap of command mode to data mode could be delayed by Zephyr
system workqueue, but MCU side is not aware of that and could
send raw data while SLM is still in command mode.

Solution is to restore old-behavior of entering data mode, that
the AT command causing SLM to enter data mode will send OK before
that, then MCU side would be sure of that SLM enters data mode.

A related change is that SLM would send below URC when quitting
data mode, signaling MCU side that SLM is back to command mode.

Signed-off-by: Jun Qing Zou <jun.qing.zou@nordicsemi.no>