applications: nrf_desktop: motion_sensor: use DEVICE_DT_GET_ONE
Use DEVICE_DT_GET_ONE to obtain an instance of a motion sensor. Note
that the architecture of the nrf_desktop could be improved to make use
of DT to further describe the configuration needed by modules. Kconfig
does this job now (e.g. to select the motion sensor), but this is likely
not the way it would be done upstream.

Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>