tfm: Add support for bootloader secondary partition in external flash
Add support for bootloader secondary partition to be placed in external
flash. This requires that TF-M does not attempt to configure this flash
partition as non-secure, since this is not supported by the SPU
peripheral.

This also fixes an issue with using the BL2 define to include the
secondary storage partition. This caused some of the structures stored
TF-M OTP secure storage to be changed, which would break during a device
firmware upgrade.

NCSDK-15345
NCSDK-15805

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>