lib: nrf_modem: nrf91_sockets: hotfix and cleanup
This commit fixes a bug introduced by a previous
commit that was setting the return value of
`getaddrinfo` to 0 even in case of error.
Also removed some error checks in `getaddrinfo`
that were not needed anymore.
Also renamed the `error` variable to `err` to
prevent confusing it with the `error` label
for `goto`.

Signed-off-by: Mirko Covizzi <mirko.covizzi@nordicsemi.no>